requirements.txt
anthropic==0.30.1
pyautogui==0.9.54
PyQt5==5.15.10
Pillow==10.4.0
pywinctl==0.2.0
pygame==2.6.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1